More slice tests & a check for StringIndexError
Starting to look like this package does the right thing, within the
strictures of "right thing" which come with using Julia.

This package wouldn't exist if there were a single obviously-correct
way for a language to handle strings, and Julia's approach is
optimal for what Julia is, and the rest of the semantics of the
language.