Removes unused findnext method
The better solution is to specify the allowed types of Pattern,
although better still would be for Base to give `i` a type in the
findnext deep within the bowels of Regex. The code is illegal to
call without an Integer and the signature makes it ambiguous to
define a method with pattern left generic.